requirements.txt
grad_cam==1.5.4
hydra-core==1.3.2
lightning==2.5.0.post0
matplotlib==3.9.4
monai==1.5.1
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
openslide-python==1.4.2
openslide-bin==4.0.0.8
optuna==4.2.1
pandas==2.2.3
pingouin==0.5.5
scikit_learn==1.5.2
scikit-image==0.24.0
scipy==1.13.1
setuptools==75.8.0
SimpleITK==2.4.1
torch==2.6.0
torchmetrics==1.7.1
torch-geometric==2.6.1
torchvision==0.21.0
psutil==7.0.0
tensorboard==2.19.0